MVVM architecture pattern applied
