+++ TravelDiscovery/Common/ActivityIndicatorView.swift
