+++ TravelDiscovery/Discover/PopularDestinations/Model/Attraction.swift
@@ -1,0 +1,14 @@
+//
+//  Attraction.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 28/06/23.
+//
+
+import Foundation
+
+struct Attraction: Identifiable {
+    let id = UUID().uuidString
+    let name, imageName: String
+    let latitude, longitude: Double
+}
