+++ TravelDiscovery/Discover/PopularDestinations/Model/DestinationsDetails.swift
@@ -1,0 +1,13 @@
+//
+//  DestinationsDetails.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 28/06/23.
+//
+
+import Foundation
+
+struct DestinationDetails: Decodable {
+    let description: String
+    let photos: [String]
+}
