+++ TravelDiscovery/Discover/PopularDestinations/View/PopularDestinationsDetailsView.swift
@@ -1,0 +1,88 @@
+//
+//  PopularDestinationsDetailsView.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 28/06/23.
+//
+
+import SwiftUI
+import MapKit
+
+struct PopularDestinationsDetailsView: View {
+    @ObservedObject var vm: DestinationDetailsViewModel
+    let destination: Destination
+    @State var region: MKCoordinateRegion
+    @State var isShowingAttractions = true
+    
+    init(destination: Destination) {
+        self.destination = destination
+        self.region = MKCoordinateRegion(center: .init(latitude: destination.latitude, longitude: destination.longitude), span: .init(latitudeDelta: 0.07, longitudeDelta: 0.07))
+        self.vm = .init(name: destination.city)
+    }
+    
+    let imagesURLString = [
+        "https://letsbuildthatapp-videos.s3-us-west-2.amazonaws.com/2240d474-2237-4cd3-9919-562cd1bb439e",
+        "https://letsbuildthatapp-videos.s3-us-west-2.amazonaws.com/b1642068-5624-41cf-83f1-3f6dff8c1702",
+        "https://letsbuildthatapp-videos.s3-us-west-2.amazonaws.com/6982cc9d-3104-4a54-98d7-45ee5d117531"
+    ]
+    
+    var body: some View {
+        ScrollView {
+            if let photos = vm.destinationDetails?.photos {
+                DestinationHeaderContainer(imageURLString: photos, isBlackBackground: false)
+                    .frame(height: 250)
+            }
+            VStack(alignment: .leading) {
+                Text(destination.city)
+                    .font(.system(size: 18, weight: .bold))
+                Text(destination.country)
+                HStack {
+                    ForEach(0..<5, id: \.self) { _ in
+                        Image(systemName: "star.fill")
+                            .foregroundColor(.orange)
+                    }
+                }
+                .padding(.top, 2)
+                
+                Text(vm.destinationDetails?.description ?? "")
+                    .padding(.top, 4)
+                    .font(.system(size: 14))
+                
+                HStack { Spacer() }
+            }
+            .padding(.horizontal)
+            HStack {
+                Text("Location")
+                    .font(.system(size: 18, weight: .semibold))
+                Spacer()
+                Button(action: {
+                    isShowingAttractions.toggle()
+                }) {
+                    Text("\(isShowingAttractions ? "Hide" : "Show") Attractions")
+                        .font(.system(size: 12, weight: .semibold))
+                }
+                Toggle("", isOn: $isShowingAttractions)
+                    .labelsHidden()
+            }
+            .padding(.horizontal)
+            Map(coordinateRegion: $region, annotationItems: isShowingAttractions ? attractions : []) { attraction in
+                MapAnnotation(coordinate: .init(latitude: attraction.latitude, longitude: attraction.longitude)) {
+                    CustomMapAnnotation(attraction: attraction)
+                }
+            }
+            .frame(height: 300)
+        }
+        .navigationBarTitle(destination.city, displayMode: .inline)
+    }
+    let attractions: [Attraction] = [
+        .init(name: "eiffel_tower", imageName: "eiffel_tower", latitude: 48.858605, longitude: 2.2946),
+        .init(name: "Champs-Elysees", imageName: "new_york", latitude: 48.866867, longitude: 2.311780),
+        .init(name: "Louvre Museum", imageName: "art2", latitude: 48.860288, longitude: 2.337789)
+    ]
+}
+
+struct PopularDestinationsDetailsView_Previews: PreviewProvider {
+    static var previews: some View {
+        PopularDestinationsDetailsView(destination: .init(city: "Paris", country: "France", imageName: "eiffel_tower", latitude: 48.859565, longitude: 2.353235))
+    }
+}
