+++ TravelDiscovery/Discover/PopularDestinations/ViewModel/DestinationDetailsViewModel.swift
@@ -1,0 +1,29 @@
+//
+//  DestinationDetailsViewModel.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 28/06/23.
+//
+
+import Foundation
+
+class DestinationDetailsViewModel: ObservableObject {
+    @Published var isLoading = true
+    @Published var destinationDetails: DestinationDetails?
+    
+    init(name: String) {
+        let urlString = "https://travel.letsbuildthatapp.com/travel_discovery/destination?name=\(name.lowercased())".addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) ?? ""
+        guard let url = URL(string: urlString ) else { return }
+        URLSession.shared.dataTask(with: url) { data, response, error in
+            DispatchQueue.main.async {
+                // make sure to check your error and response
+                guard let data = data else { return }
+                do {
+                    self.destinationDetails = try JSONDecoder().decode(DestinationDetails.self, from: data)
+                } catch {
+                    print("Failed to decode JSON", error)
+                }
+            }
+        }.resume()
+    }
+}
